implement scattering(self, x) in 1D base frontend
