+++ kymatio/scattering1d/frontend/tensorflow_frontend.py
@@ -17,30 +17,6 @@ class ScatteringTensorFlow1D(ScatteringTensorFlow, ScatteringBase1D):
         ScatteringBase1D.build(self)
         ScatteringBase1D.create_filters(self)
 
-    def scattering(self, x):
-        ScatteringBase1D._check_runtime_args(self)
-        ScatteringBase1D._check_input(self, x)
-        x_shape = self.backend.shape(x)
-        batch_shape, signal_shape = x_shape[:-1], x_shape[-1:]
-        x = self.backend.reshape_input(x, signal_shape, n_inserted_dims=1)
-
-        S = scattering1d(x, self.backend, self.psi1_f, self.psi2_f,
-                         self.phi_f, max_order=self.max_order, average=self.average, pad_left=self.pad_left,
-                         pad_right=self.pad_right, ind_start=self.ind_start, ind_end=self.ind_end,
-                         oversampling=self.oversampling)
-
-        n_kept_dims = 1 + (self.out_type=="dict")
-        for n, path in enumerate(S):
-            S[n]['coef'] = self.backend.reshape_output(
-                path['coef'], batch_shape, n_kept_dims=n_kept_dims)
-
-        if self.out_type=='array':
-            return self.backend.concatenate([path['coef'] for path in S], dim=-2)
-        elif self.out_type=='dict':
-            return {path['n']: path['coef'] for path in S}
-        elif self.out_type=='list':
-            return list(map(lambda path: path.pop('n')), S)
-
 
 ScatteringTensorFlow1D._document()
